Extract TensorYlm into its own library
Reduces the size of the low-level SphericalHarmonics library
and avoids cyclic dependence on Spectral.
