+++ src/Evolution/Systems/GeneralizedHarmonic/CMakeLists.txt
@@ -61,6 +61,7 @@ target_link_libraries(
   Serialization
   Spectral
   SphericalHarmonics
+  TensorYlm
   Utilities
   INTERFACE
   Initialization
