+++ src/Evolution/Systems/GrMhd/GhValenciaDivClean/ApplyTensorYlmFilter.cpp
@@ -13,12 +13,12 @@
 #include "Evolution/Systems/RadiationTransport/NoNeutrinos/System.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/Spherepack.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/SpherepackCache.hpp"
-#include "NumericalAlgorithms/SphericalHarmonics/TensorYlmFilter.hpp"
+#include "NumericalAlgorithms/TensorYlm/Filter.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/TMPL.hpp"
 
-#include "NumericalAlgorithms/SphericalHarmonics/ApplyTensorYlmFilter.tpp"
+#include "NumericalAlgorithms/TensorYlm/ApplyFilter.tpp"
 
 namespace ylm::TensorYlm {
 template <>
