+++ src/NumericalAlgorithms/CMakeLists.txt
@@ -15,3 +15,4 @@ add_subdirectory(SpatialDiscretization)
 add_subdirectory(Spectral)
 add_subdirectory(SphericalHarmonics)
 add_subdirectory(SpinWeightedSphericalHarmonics)
+add_subdirectory(TensorYlm)
