+++ src/NumericalAlgorithms/TensorYlm/CartToSphere.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "NumericalAlgorithms/SphericalHarmonics/TensorYlm.hpp"
+#include "NumericalAlgorithms/TensorYlm/TensorYlm.hpp"
 
 #include <cstddef>
