+++ tests/Unit/NumericalAlgorithms/CMakeLists.txt
@@ -14,3 +14,4 @@ add_subdirectory(RootFinding)
 add_subdirectory(Spectral)
 add_subdirectory(SphericalHarmonics)
 add_subdirectory(SpinWeightedSphericalHarmonics)
+add_subdirectory(TensorYlm)
