+++ tests/Unit/NumericalAlgorithms/SphericalHarmonics/CMakeLists.txt
@@ -13,10 +13,6 @@ set(LIBRARY_SOURCES
   Test_StrahlkorperDataBox.cpp
   Test_StrahlkorperFunctions.cpp
   Test_Tags.cpp
-  Test_TensorYlmCartToSphere.cpp
-  Test_TensorYlmFilter.cpp
-  Test_TensorYlmSphereToCart.cpp
-  Test_WignerThreeJ.cpp
   Test_YlmToStf.cpp
   )
